+++ src/unit_tests/variant/variant_conv_to_int64.cpp
@@ -263,7 +263,7 @@ TEST_CASE("variant::to_int64() - from int8_t", "[variant]")
     {
         variant var = int8_t(-60);
         bool ok = false;
-        CHECK(var.to_int64(&ok) == int8_t(-60));
+        CHECK(var.to_int64(&ok) == int64_t(-60));
         CHECK(ok == true);
         CHECK(var.convert(type::get<int64_t>()) == true);
     }
@@ -278,18 +278,18 @@ TEST_CASE("variant::to_int64() - from int16_t", "[variant]")
         variant var = int16_t(32760);
         REQUIRE(var.can_convert<int64_t>() == true);
         bool ok = false;
-        CHECK(var.to_int64(&ok) == int16_t(32760));
+        CHECK(var.to_int64(&ok) == int64_t(32760));
         CHECK(ok == true);
 
         CHECK(var.convert(type::get<int64_t>()) == true);
-        CHECK(var.get_value<int64_t>() == int16_t(32760));
+        CHECK(var.get_value<int64_t>() == int64_t(32760));
     }
 
     SECTION("valid conversion negative")
     {
         variant var = int16_t(-32760);
         bool ok = false;
-        CHECK(var.to_int64(&ok) == int16_t(-32760));
+        CHECK(var.to_int64(&ok) == int64_t(-32760));
         CHECK(ok == true);
         CHECK(var.convert(type::get<int64_t>()) == true);
     }
@@ -304,18 +304,18 @@ TEST_CASE("variant::to_int64() - from int32_t", "[variant]")
         variant var = int32_t(2147483640);
         REQUIRE(var.can_convert<int64_t>() == true);
         bool ok = false;
-        CHECK(var.to_int64(&ok) == int32_t(2147483640));
+        CHECK(var.to_int64(&ok) == int64_t(2147483640));
         CHECK(ok == true);
 
         CHECK(var.convert(type::get<int64_t>()) == true);
-        CHECK(var.get_value<int64_t>() == int32_t(2147483640));
+        CHECK(var.get_value<int64_t>() == int64_t(2147483640));
     }
 
     SECTION("valid conversion negative")
     {
         variant var = int32_t(-2147483640);
         bool ok = false;
-        CHECK(var.to_int64(&ok) == int32_t(-2147483640));
+        CHECK(var.to_int64(&ok) == int64_t(-2147483640));
         CHECK(ok == true);
         CHECK(var.convert(type::get<int64_t>()) == true);
     }
@@ -359,8 +359,8 @@ TEST_CASE("variant::to_int64() - from uint8_t", "[variant]")
         CHECK(var.to_int64(&ok) == 50);
         CHECK(ok == true);
 
-        CHECK(var.convert(type::get<uint8_t>()) == true);
-        CHECK(var.get_value<uint8_t>() == 50);
+        CHECK(var.convert(type::get<int64_t>()) == true);
+        CHECK(var.get_value<int64_t>() == int64_t(50));
     }
 }
 
@@ -373,11 +373,11 @@ TEST_CASE("variant::to_int64() - from uint16_t", "[variant]")
         variant var = uint16_t(32760);
         REQUIRE(var.can_convert<uint16_t>() == true);
         bool ok = false;
-        CHECK(var.to_int64(&ok) == uint16_t(32760));
+        CHECK(var.to_int64(&ok) == int64_t(32760));
         CHECK(ok == true);
 
-        CHECK(var.convert(type::get<uint16_t>()) == true);
-        CHECK(var.get_value<uint16_t>() == uint16_t(32760));
+        CHECK(var.convert(type::get<int64_t>()) == true);
+        CHECK(var.get_value<int64_t>() == int64_t(32760));
     }
 }
 
@@ -390,11 +390,11 @@ TEST_CASE("variant::to_int64() - from uint32_t", "[variant]")
         variant var = uint32_t(32760);
         REQUIRE(var.can_convert<uint32_t>() == true);
         bool ok = false;
-        CHECK(var.to_int64(&ok) == uint32_t(32760));
+        CHECK(var.to_int64(&ok) == int64_t(32760));
         CHECK(ok == true);
 
-        CHECK(var.convert(type::get<uint32_t>()) == true);
-        CHECK(var.get_value<uint32_t>() == uint32_t(32760));
+        CHECK(var.convert(type::get<int64_t>()) == true);
+        CHECK(var.get_value<int64_t>() == int64_t(32760));
     }
 }
 
@@ -407,11 +407,11 @@ TEST_CASE("variant::to_int64() - from uint64_t", "[variant]")
         variant var = uint64_t(2147483640);
         REQUIRE(var.can_convert<uint64_t>() == true);
         bool ok = false;
-        CHECK(var.to_int64(&ok) == uint64_t(2147483640));
+        CHECK(var.to_int64(&ok) == int64_t(2147483640));
         CHECK(ok == true);
 
-        CHECK(var.convert(type::get<uint64_t>()) == true);
-        CHECK(var.get_value<uint64_t>() == uint64_t(2147483640));
+        CHECK(var.convert(type::get<int64_t>()) == true);
+        CHECK(var.get_value<int64_t>() == int64_t(2147483640));
     }
 
     SECTION("too big")
